Feat/back/#76 event crud

+++ src/main/java/com/happiday/Happi_Day/domain/service/EventService.java
@@ -5,7 +5,10 @@ import com.happiday.Happi_Day.domain.entity.event.dto.EventListResponseDto;
 import com.happiday.Happi_Day.domain.entity.event.dto.EventResponseDto;
 import com.happiday.Happi_Day.domain.entity.event.dto.EventUpdateDto;
 import com.happiday.Happi_Day.domain.entity.event.Event;
+import com.happiday.Happi_Day.domain.entity.user.User;
 import com.happiday.Happi_Day.domain.repository.EventRepository;
+import com.happiday.Happi_Day.domain.repository.UserRepository;
+import com.happiday.Happi_Day.utils.FileUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -13,9 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,38 +24,60 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class EventService {
     private final EventRepository eventRepository;
-
-// TODO    private final ArtistRepository artistRepository;
+    private final UserRepository userRepository;
+    private final FileUtils fileUtils;
 
     @Transactional
-    public EventResponseDto createEvent(EventCreateDto request, List<MultipartFile> eventImages) {
-        // TODO - 유저, 이미지 추가
+    public EventResponseDto createEvent(
+            EventCreateDto request, MultipartFile thumbnailFile, MultipartFile imageFile, String username) {
 
-        Event event = request.toEntity();
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        // 이미지 로직 추가
+        String thumbnailUrl = fileUtils.uploadFile(thumbnailFile);
+        String imageUrl = fileUtils.uploadFile(imageFile);
+
+        Event event = request.toEntity(thumbnailUrl, imageUrl);
         eventRepository.save(event);
         return EventResponseDto.fromEntity(event);
     }
 
     public  List<EventListResponseDto> readEvents() {
-        // TODO - 이미지 추가
         return eventRepository.findAll().stream()
                 .map(EventListResponseDto::fromEntity)
                 .collect(Collectors.toList());
     }
 
     public EventResponseDto readEvent(Long eventId) {
-        // TODO - 이미지 추가
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
         return EventResponseDto.fromEntity(event);
     }
 
     @Transactional
-    public EventResponseDto updateEvent(Long eventId, EventUpdateDto request, List<MultipartFile> eventImages) {
-        // TODO - 유저, 이미지 추가
+    public EventResponseDto updateEvent(
+            Long eventId, EventUpdateDto request, MultipartFile thumbnailFile, MultipartFile imageFile, String username) {
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        if(!user.getUsername().equals(username)) throw new IllegalArgumentException("사용자 정보가 일치하지 않습니다.");
 
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        // 이미지 업로드 추가
+        if (thumbnailFile != null && !thumbnailFile.isEmpty()) {
+            fileUtils.deleteFile(event.getThumbnailUrl());
+            String newThumbnailUrl = fileUtils.uploadFile(thumbnailFile);
+            event.setThumbnailUrl(newThumbnailUrl);
+        }
+
+        if (imageFile != null && !imageFile.isEmpty()) {
+            fileUtils.deleteFile(event.getImageUrl());
+            String newImageUrl = fileUtils.uploadFile(imageFile);
+            event.setImageUrl(newImageUrl);
+        }
 
         event.update(request.toEntity());
         eventRepository.save(event);
@@ -62,7 +85,12 @@ public class EventService {
     }
 
     @Transactional
-    public void deleteEvent(Long eventId) {
+    public void deleteEvent(Long eventId, String username) {
+        User user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        if(!user.getUsername().equals(username)) throw new IllegalArgumentException("사용자 정보가 일치하지 않습니다.");
+
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
         eventRepository.delete(event);
